requirements.txt
torch
transformers
PyYAML
tqdm
numpy